requirements.txt
keyboard==0.13.5
mss==10.0.0
pyautogui==0.9.54
pytesseract==0.3.13
pywin32==311
XlsxWriter==3.2.2

PySide6~=6.10.0
setuptools~=78.1.0
numpy~=2.3.0  # 或 numpy~=2.4.0
pillow~=12.1.0
opencv-python~=4.11.0.86
pandas~=3.0.0
openpyxl~=3.1.5